Change id to Long variables

+++ backend/src/main/java/com/software/backend/controllers/dtos/PacienteDTO.java
@@ -11,10 +11,10 @@ import jakarta.validation.constraints.Size;
 public class PacienteDTO {
 
     @NotNull(message = "El CUIL no puede ser nulo")
-    private Integer cuil;
+    private long cuil;
 
     @NotNull(message = "El DNI no puede ser nulo")
-    private Integer dni;
+    private long dni;
 
     @NotNull(message = "La fecha de nacimiento no puede ser nula")
     private Date fechaNacimiento;
@@ -25,7 +25,7 @@ public class PacienteDTO {
 
     @NotNull(message = "El número de teléfono no puede ser nulo")
     @Digits(integer = 15, fraction = 0, message = "El número de teléfono debe ser válido")
-    private Integer telefono;
+    private long telefono;
 
     @NotBlank(message = "El nombre no puede estar vacío")
     @Size(min = 2, max = 255, message = "El nombre debe tener entre 2 y 255 caracteres")
@@ -42,11 +42,11 @@ public class PacienteDTO {
     
     private ObraSocialDTO obraSocial;
 
-    public PacienteDTO(@NotNull(message = "El CUIL no puede ser nulo") Integer cuil,
-            @NotNull(message = "El DNI no puede ser nulo") Integer dni,
+    public PacienteDTO(@NotNull(message = "El CUIL no puede ser nulo") long cuil,
+            @NotNull(message = "El DNI no puede ser nulo") long dni,
             @NotNull(message = "La fecha de nacimiento no puede ser nula") Date fechaNacimiento,
             @Email(message = "El email debe ser válido") @Size(min = 2, max = 255, message = "El apellido debe tener entre 2 y 255 caracteres") String email,
-            @NotNull(message = "El número de teléfono no puede ser nulo") @Digits(integer = 15, fraction = 0, message = "El número de teléfono debe ser válido") Integer telefono,
+            @NotNull(message = "El número de teléfono no puede ser nulo") @Digits(integer = 15, fraction = 0, message = "El número de teléfono debe ser válido") long telefono,
             @NotBlank(message = "El nombre no puede estar vacío") @Size(min = 2, max = 255, message = "El nombre debe tener entre 2 y 255 caracteres") String nombre,
             @NotBlank(message = "El apellido no puede estar vacío") @Size(min = 2, max = 255, message = "El apellido debe tener entre 2 y 255 caracteres") String apellido,
             @NotNull(message = "El número de afiliado no puede ser nulo") Integer numeroAfiliado,
@@ -62,19 +62,19 @@ public class PacienteDTO {
         this.obraSocial = obraSocial;
     }
 
-    public Integer getCuil() {
+    public long getCuil() {
         return cuil;
     }
 
-    public void setCuil(Integer cuil) {
+    public void setCuil(long cuil) {
         this.cuil = cuil;
     }
 
-    public Integer getDni() {
+    public long getDni() {
         return dni;
     }
 
-    public void setDni(Integer dni) {
+    public void setDni(long dni) {
         this.dni = dni;
     }
 
@@ -94,11 +94,11 @@ public class PacienteDTO {
         this.email = email;
     }
 
-    public Integer getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(Integer telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
